+++ src/cudadecoder/Makefile
@@ -18,7 +18,7 @@ OBJFILES = cuda-decoder.o cuda-decoder-kernels.o cuda-fst.o \
 	   batched-threaded-nnet3-cuda-pipeline.o \
 	   batched-threaded-nnet3-cuda-pipeline2.o \
 	   batched-static-nnet3.o batched-static-nnet3-kernels.o \
-	   deprecated/decodable-cumatrix.o
+	   decodable-cumatrix.o
 
 LIBNAME = kaldi-cudadecoder
