+++ src/cudadecoder/batched-threaded-nnet3-cuda-pipeline.h
@@ -22,8 +22,8 @@
 #include <thread>
 
 #include "cudadecoder/cuda-decoder.h"
-#include "cudadecoder/deprecated/decodable-cumatrix.h"
-#include "cudadecoder/deprecated/thread-pool.h"
+#include "cudadecoder/decodable-cumatrix.h"
+#include "cudadecoder/thread-pool.h"
 #include "cudafeat/online-cuda-feature-pipeline.h"
 #include "feat/wave-reader.h"
 #include "lat/determinize-lattice-pruned.h"
